backend/http-server/requirements.txt
grpcio==1.32.0
grpcio-tools==1.32.0
tornado==6.0.4
pydevd-pycharm~=201.7846.77